requirements.txt
pdfplumber
python-docx
requests
sentence-transformers
faiss-cpu
google-generativeai
ollama>=0.3.0